refactor: extract common response fields
